feat. 844280 Add mechanism for SDK to pass M symbols to vsc-pq p5
